requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
Naked==0.1.32
paytmchecksum==1.7.0
pillow==11.3.0
pycryptodome==3.23.0
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
shellescape==3.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
